Created eclipse project files
