+++ src/DessinBaigneurs.java
@@ -1,0 +1,13 @@
+import javax.swing.JFrame;
+
+public class DessinBaigneurs {
+
+	public static void main(String[] args){
+		
+		JFrame fenetre = new JFrame();
+		fenetre.setBounds(100,100,800,400);
+		fenetre.setVisible(true);
+		
+	}
+	
+}
