--- src/devoir_se/DessinBaigneurs.java
@@ -1,5 +1,0 @@
-package devoir_se;
-
-public class DessinBaigneurs {
-
-}
